# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Affiliate Backend Application",
        "title": "Affiliate Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/advertisers": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new advertiser with the given details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisers"
                ],
                "summary": "Create a new advertiser",
                "parameters": [
                    {
                        "description": "Advertiser details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateAdvertiserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created advertiser",
                        "schema": {
                            "$ref": "#/definitions/domain.Advertiser"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/advertisers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves an advertiser by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisers"
                ],
                "summary": "Get advertiser by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Advertiser ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Advertiser details",
                        "schema": {
                            "$ref": "#/definitions/domain.Advertiser"
                        }
                    },
                    "400": {
                        "description": "Invalid advertiser ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Advertiser not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an advertiser with the given details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisers"
                ],
                "summary": "Update advertiser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Advertiser ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Advertiser details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateAdvertiserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated advertiser",
                        "schema": {
                            "$ref": "#/definitions/domain.Advertiser"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Advertiser not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an advertiser by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisers"
                ],
                "summary": "Delete advertiser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Advertiser ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Invalid advertiser ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Advertiser not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/organizations/{id}/advertisers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of advertisers for an organization with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisers"
                ],
                "summary": "List advertisers by organization",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Organization ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default: 10)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of advertisers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Advertiser"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid organization ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Advertiser": {
            "type": "object",
            "properties": {
                "advertiser_id": {
                    "type": "integer"
                },
                "billing_details": {
                    "description": "JSONB stored as string",
                    "type": "string"
                },
                "contact_email": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "integer"
                },
                "status": {
                    "description": "'active', 'pending', 'inactive', 'rejected'",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CreateAdvertiserRequest": {
            "type": "object",
            "required": [
                "name",
                "organization_id"
            ],
            "properties": {
                "billing_details": {
                    "description": "Billing details in JSON format\nswagger:strfmt json",
                    "type": "object"
                },
                "contact_email": {
                    "description": "Contact email address",
                    "type": "string",
                    "example": "contact@example.com"
                },
                "name": {
                    "description": "Advertiser name",
                    "type": "string",
                    "example": "Example Advertiser"
                },
                "organization_id": {
                    "description": "Organization ID",
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "description": "Status of the advertiser (active, pending, inactive, rejected)",
                    "type": "string",
                    "example": "active"
                }
            }
        },
...
        "handlers.UpdateAdvertiserRequest": {
            "type": "object",
            "required": [
                "name",
                "status"
            ],
            "properties": {
                "billing_details": {
                    "description": "Billing details in JSON format\nswagger:strfmt json",
                    "type": "object"
                },
                "contact_email": {
                    "description": "Contact email address",
                    "type": "string",
                    "example": "updated@example.com"
                },
                "name": {
                    "description": "Advertiser name",
                    "type": "string",
                    "example": "Updated Advertiser"
                },
                "status": {
                    "description": "Status of the advertiser (active, pending, inactive, rejected)",
                    "type": "string",
                    "example": "active"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}